backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.1",
    "fastapi[standard]>=0.121.3",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "supabase>=2.24.0",
    "uvicorn>=0.38.0",
]
